+++ Scripts/WoW Game Data API/Pet API/Endpoint_Wow_GameData_PetAbilitiesIndex.cs
@@ -1,0 +1,70 @@
+// ╔════════════════════════════════════╗
+// ║ This file has been auto-generated. ║
+// ╚════════════════════════════════════╝
+
+using System;
+using System.Collections;
+using ZeShmouttsAssets.BlizzardAPI.JSON;
+
+namespace ZeShmouttsAssets.BlizzardAPI
+{
+	/// <summary>
+	/// Interface for working with the Blizzard API inside Unity.
+	/// </summary>
+	public static partial class BlizzardAPI
+	{
+		/// <summary>
+		/// API endpoints related to World of Warcraft game data (items, spells, etc.).
+		/// Reference : https://develop.battle.net/documentation/world-of-warcraft/game-data-apis
+		/// </summary>
+		public static partial class WowGameData
+		{
+			internal const string apiPath_PetAbilitiesIndex = basePath_Wow_gameData + "/pet-ability/index";
+
+			/// <summary>
+			/// Coroutine that retrieves an index of pet abilities.
+			/// </summary>
+			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetPetAbilitiesIndex(Action<Json_Wow_PetAbilitiesIndex> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = apiPath_PetAbilitiesIndex;
+				yield return SendRequest(region, namespaceStatic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+			/// <summary>
+			/// Coroutine that retrieves an index of pet abilities, as a raw JSON string.
+			/// </summary>
+			/// <param name="action_Result">Action to execute with the raw JSON string.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetPetAbilitiesIndexRaw(Action<string> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = apiPath_PetAbilitiesIndex;
+				yield return SendRequest(region, namespaceStatic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+		}
+	}
+}
+
+namespace ZeShmouttsAssets.BlizzardAPI.JSON
+{
+	/// <summary>
+	/// JSON structure for World of Warcraft, representing an index of pet abilities.
+	/// </summary>
+	[Serializable]
+	public class Json_Wow_PetAbilitiesIndex : Object_JSON
+	{
+		// {{JSON_START}}
+		public LinkStruct _links;
+
+		public RefNameIdStruct[] abilities;
+		// {{JSON_END}}
+	}
+}